+++ CloudShell.Configuration.Client/CloudShellConfigurationSetting.cs
@@ -1,0 +1,5 @@
+namespace CloudShell.Configuration.Client;
+
+public sealed record CloudShellConfigurationSetting(
+    string Name,
+    string Value);